integer overflow in tensorid leading to crash

+++ habana_lazy/hpu_lazy_tensors.h
@@ -228,7 +228,7 @@ class HbLazyTensor {
 // operations and ensure the same computations are created during the
 // training loops.
 struct HbContext {
-  std::map<int, std::weak_ptr<Data>> tensors_data;
+  std::map<int64_t, std::weak_ptr<Data>> tensors_data;
   ir::Value seed_ir_value;
 };
